First shot at retrieving complex correspondences
We can now ask the ontology matcher to provide correspondences for
entire patterns. It therefore slices the pattern into increasingly
smaller chunks and tries to retrieve matches for each of them.

What is still missing:
- what should we do with multiple matches (i.e., correspondence for
entire pattern + correspondences for subgraphs)
- the query still return partial matches, i.e., if there is additional
stuff in the correspondence, it gets returned, as well -> figure our a
way to avoid that or at least make it predictable